Updated main.c of Tiva
The Tiva was unable to receive messages from the VCUs because the "MSG_OBJ_EXTENDED_ID " flag was not configured to be used when receiving a message into the respective message box - the VCUs were using Extended ID  with their messages and the Tiva node was expecting a standard ID to filter messages from the bus.

Currently:  Tiva can now talk to each Individual VCUs, consistently. However, the Tiva and the 2 VCUs on the same network, are not consistent with sending and receiving messages.